requirements.txt
crewai==0.102.0
tavily-python
langchain
python-dotenv
streamlit